Move all integration test to a separate test command
