+++ package.json
@@ -10,6 +10,7 @@
     "start:prod": "node dist/main",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "jest",
+    "test:integration": "jest --config=src/tests/integration/jest.integration.config.ts",
     "debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "migrate": "sequelize db:migrate",
     "migrate:create": "sequelize db:create",
